Renamed UUID resolver to ref resolver
